# Optimize MTrie reading single path by adding Ledger.GetSingleValue() to boost speed and reduce memory use