--- a/ledger/trie.go
+++ b/ledger/trie.go
@@ -89,6 +89,12 @@ type TrieRead struct {
 	Paths    []Path
 }
 
+// TrieReadSinglePayload contains trie read query for a single payload
+type TrieReadSinglePayload struct {
+	RootHash RootHash
+	Path     Path
+}
+
 // TrieUpdate holds all data for a trie update
 type TrieUpdate struct {
 	RootHash RootHash