### Changes

When reading a payload from a single path, avoid needlessly path partitioning and recursive calls.  Also, optimize MTrie batch read.

- Added Ledger.GetSingleValue() to improve time/op, alloc/op, and allocs/op
- Modified delta.View.readFunc to use Ledger.GetSingleValue()
- Optimized existing MTrie batch read for single path

Closes #1876 

### Benchmark comparisions

####  Current Ledger.Get() vs Ledger.GetSingleValue()

```
Benchstat results:
name                           old time/op    new time/op    delta
LedgerGetOneValue/batch_get-4    6.54µs ± 0%    5.24µs ± 0%  -19.92%
(p=0.000 n=9+10)

name                           old alloc/op   new alloc/op   delta
LedgerGetOneValue/batch_get-4    1.74kB ± 0%    1.62kB ± 0%   -6.85%
(p=0.000 n=10+10)

name                           old allocs/op  new allocs/op  delta
LedgerGetOneValue/batch_get-4      21.0 ± 0%      15.0 ± 0%  -28.57%
(p=0.000 n=10+10)
```

#### Optimized new Ledger.Get() vs Ledger.GetSingleValue()

```
Benchstat results:
name                           old time/op    new time/op    delta
LedgerGetOneValue/batch_get-4    5.70µs ± 0%    5.23µs ± 0%   -8.24%
(p=0.000 n=9+10)

name                           old alloc/op   new alloc/op   delta
LedgerGetOneValue/batch_get-4    1.69kB ± 0%    1.62kB ± 0%   -3.79%
(p=0.000 n=10+10)

name                           old allocs/op  new allocs/op  delta
LedgerGetOneValue/batch_get-4      18.0 ± 0%      15.0 ± 0%  -16.67%
(p=0.000 n=10+10)
```